move mapped cemu controllers to bak if they exist

+++ functions/EmuScripts/emuDeckCemu.sh
@@ -55,9 +55,16 @@ Cemu_init(){
 	Cemu_setEmulationFolder
 	Cemu_setupSaves
 	Cemu_addSteamInputProfile
-	mv "${romsPath}/wiiu/controllerProfiles/controller1.xml" "${romsPath}/wiiu/controllerProfiles/controller1.xml.bak"
-	mv "${romsPath}/wiiu/controllerProfiles/controller2.xml" "${romsPath}/wiiu/controllerProfiles/controller2.xml.bak"
-	mv "${romsPath}/wiiu/controllerProfiles/controller3.xml" "${romsPath}/wiiu/controllerProfiles/controller3.xml.bak"
+
+	if [ -e "${romsPath}/wiiu/controllerProfiles/controller1.xml" ];then 
+		mv "${romsPath}/wiiu/controllerProfiles/controller1.xml" "${romsPath}/wiiu/controllerProfiles/controller1.xml.bak"
+	fi
+	if [ -e "${romsPath}/wiiu/controllerProfiles/controller2.xml" ];then 
+		mv "${romsPath}/wiiu/controllerProfiles/controller2.xml" "${romsPath}/wiiu/controllerProfiles/controller2.xml.bak"
+	fi
+	if [ -e "${romsPath}/wiiu/controllerProfiles/controller3.xml" ];then 
+		mv "${romsPath}/wiiu/controllerProfiles/controller3.xml" "${romsPath}/wiiu/controllerProfiles/controller3.xml.bak"
+	fi
 }
 
 #update
